Fix -Wswitch/-Wshorten-64-to-32/-Wlogical-op-parentheses warnings
Reported from an openFrameworks apothecary CI build
(github.com/openframeworks/apothecary/actions/runs/32219237031).
Reproduced locally by building with the same warning flags enabled
(-Wswitch -Wshorten-64-to-32 -Wlogical-op-parentheses
-Wlogical-not-parentheses) and fixed everything in scope (Source/
FreeImage, FreeImageToolkit, Metadata - not the bundled third-party
libs, which weren't part of the reported warnings).

-Wswitch (missing enum cases): added `default: break;` to switches
that intentionally only handle a subset of FREE_IMAGE_TYPE/
FREE_IMAGE_COLOR_CHANNEL/etc values. Purely mechanical/behavior-
preserving, but surfaced one real pre-existing bug it exposed rather
than caused: Rescale.cpp's `case FIT_RGBAF` declared a local
(`FIBITMAP *rgbf`) without a block scope, which becomes a hard
compile error ("cannot jump to case label") once anything follows it
in the switch - wrapped that case body in braces, the correct fix.

-Wshorten-64-to-32: added explicit casts at ~25 sites (file offset
arithmetic via tell_proc/seek_proc returning `long`, zlib's
uLong/uLongf, strtol results) assigned into narrower DWORD/unsigned/
int fields. All are already implicitly narrowed today (this is
long-standing, working behavior on LLP64/LP64 platforms) - just
making the intentional narrowing visible to the compiler, not
changing any behavior.

-Wlogical-op-parentheses / -Wlogical-not-parentheses: reviewed each
of the 7 sites individually rather than blindly adding parens, since
`&&` binding tighter than `||` can silently hide a real logic bug.
Six were confirmed correct as originally written (verified the
De Morgan's-equivalent intent against surrounding comments/context)
and just needed explicit parens for clarity. One was a genuine bug:
PluginTIFF.cpp's IsValidBitsPerSample() checked
`(photometric == PHOTOMETRIC_RGB) && (samplesperpixel == 3) ||
(samplesperpixel == 4)` for its "// RGB[A]F" 32-bit-per-sample
validation - as written, any 4-samples-per-pixel image (e.g. CMYK)
would incorrectly validate as RGBA float, regardless of photometric.
Fixed to `(photometric == PHOTOMETRIC_RGB) && ((samplesperpixel == 3)
|| (samplesperpixel == 4))`, matching the comment's clear intent.
Also simplified a `-Wlogical-not-parentheses` site in PluginPICT.cpp
(`!x == 0`) to the unambiguous `x != 0` - mathematically equivalent
either way `!x == 0` is parsed, but confusing to read.

Verified: rebuilding with all four warning flags now produces zero
warnings in the affected files, the full test suite passes, and a
plain default-flags build (matching normal CI) also builds clean.

+++ Source/FreeImage/ColorLookup.cpp
@@ -574,7 +574,7 @@ FreeImage_LookupX11Color(const char *szColor, BYTE *nRed, BYTE *nGreen, BYTE *nB
          (szColor[3] == 'y' || szColor[3] == 'Y' ) )  {
 
         // grey<num>, or gray<num>, num 1...100
-        i = strtol(szColor+4, NULL, 10);
+        i = (int)strtol(szColor+4, NULL, 10);
         *nRed   = (BYTE)(255.0/100.0 * i);
         *nGreen = *nRed;
         *nBlue  = *nRed;
@@ -768,7 +768,7 @@ FreeImage_LookupSVGColor(const char *szColor, BYTE *nRed, BYTE *nGreen, BYTE *nB
          (szColor[3] == 'y' || szColor[3] == 'Y' ) )  {
 
         // grey<num>, or gray<num>, num 1...100
-        i = strtol(szColor+4, NULL, 10);
+        i = (int)strtol(szColor+4, NULL, 10);
         *nRed   = (BYTE)(255.0/100.0 * i);
         *nGreen = *nRed;
         *nBlue  = *nRed;
